New DebugUI plugin
This plugin is meant to have some add-ons to interact with the
simulation to provide insight.

This first change adds a cursor under the mouse with the intention that
it's circle can be used as a reticule for targeting boids. Hovering over
a cluster and getting their CoM or velocities, for example.