Pin transitive dep ruzstd=0.8.1
The semver-compatible ruzstd 0.8.2 uses a feature that Rust 1.85 thinks
is "nightly only." This might have been true when 1.85 was released,
but it must have since been stabilized.

I have to patch this for the Debian package to build correctly, and it
needs to happen here because I need to vendor the dependency.